agents/vulnerable-app/app/requirements.txt
# Vulnerable App Dependencies
# NHI Security Testbed

flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
